Create population.c
A program that calculates how many years will take for population to grow.

The program asks for user inputs of start population and end population. Minimum start population needs to be 9 or higher. if every year there is starts / 3 newborns and start / 4 deceased, how many years will take for the population to grow to the end number. The increase caused by newborns each year is taken into account. 
Conditions are used so the program will prompt for an input again, if the user set start population < 9 or end population < start population.

For example:
Start population: 9
End population: 19
Years: 8

So the program calculates that 8 years will be needed for the population to grow from 9 to 18 individuals.